service-crud Refactor resolveRouteBinding into trait
