+++ application/app/Models/Customer.php
@@ -3,7 +3,7 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Builder;
+use App\Traits\ScopedRouteModelBindingTrait;
 use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -16,6 +16,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 class Customer extends Model
 {
     use HasFactory;
+    use ScopedRouteModelBindingTrait;
 
     protected $fillable = [
         'user_id',
@@ -23,22 +24,6 @@ class Customer extends Model
         'tax_number',
         'tax_rate',
     ];
-
-    /**
-     * This ensures the route model binding always loads the record owned by the logged-in user
-     *
-     * @param $value
-     * @param $field
-     * @return Model|Builder
-     */
-    public function resolveRouteBinding($value, $field = null): Model|Builder
-    {
-        return $this
-            ->query()
-            ->where('id', $value)
-            ->where('user_id', auth()->id())
-            ->firstOrFail();
-    }
 
     public function user(): BelongsTo
     {
